refactor(worker): extract step reporting from job execution
